Speed improvement to lw_expr
Instead of using a singly linked list and doing the slow append algorithm
when adding an operand to an expression (scan from the start to find the
end), now maintain a tail pointer. Also maintain a previous pointer in each
entry. Benchmarking suggests this yields a rougly 30% improvement in
runtime.

Also refactor some of the code in lw_expr.c to make it somewhat clearer to
understand. For some values of clearer and understand.