Make base prefix sigils error out if no number following
Prefix sigils (0x, %, &, @, $) were parsing to 0 if there was no number
following the prefix. Make parsing fail if that is the case. Note that in a
couple of obscure cases, it may give "undefined symbol" rather than "bad
operand" due to some interactions with other unfortunate features of the
source format.