Fix some signedness mismatches
Use size_t instead of int for stuff that's going to be compared with the
strlen() return value. While not technically wrong to use int, it does seem
to attract warnings from some compilers. Technically, since it's also the
return value from strtol(), there should be one variable that is long and
another that is size_t but that's just getting crazy.

Thanks to Erik G <erik@6809.org> for pointing these ones out.