Port remainder of tools to Windows
Thanks to Erik G <erik@6809.org> for the patch.

Quoth Erik:

"As a thank you for putting up with my nonsense, I ported the remaining
binaries. Although I'm not using them (yet), I did verify that they work and
the changes are basically zero so I have confidence in them."