Completed initial conversion to new parser allowing spaces in operands
Converted the remaining addressing modes. This required a complete rewrite
of a large portion of the indexed addressing parser. Now the entire indexed
parsing system is programmatic without cheating with a lookup table.

This update also fixes the "force 0,r" by writing a literal 0,r which is
*supposed* to work.

There will likely be some pseudo ops that need tweaking for space handling,
specially those that take multiple operands of some description which are
not expressions. (The expression parser call eats the spaces both before and
after the expression, if appropriate.)