Fix cycle counting for relative branches.
The original creator of the cycle counting code appears to have mangled two
different approaches to counting the cycles for relative branches into the
final product which led to counting too many cycles for the branch
instructions. The comments in the cycle table were correct but the code was
doing some incorrect things.

Also add the missing LBRN entry to the cycle table. The one labelled LBRN
was not, in fact, LBRN, but one of the actual useful branch instructions.